oh-tab-kdl1: split InputArea popovers/components
