+++ src/webview-src/components/InputArea.tsx
@@ -1,5 +1,10 @@
 import { useRef, useState, useEffect } from 'react';
-import { useCloseOnEscapeAndOutsideClick, type CloseReason } from './useCloseOnEscapeAndOutsideClick';
+import type { CloseReason } from './useCloseOnEscapeAndOutsideClick';
+import { AccessoryButton } from './inputArea/AccessoryButton';
+import { ContextPicker } from './inputArea/ContextPicker';
+import { LlmProfileSelector } from './inputArea/LlmProfileSelector';
+import { SkillsPopover } from './inputArea/SkillsPopover';
+import { ToolsPopover } from './inputArea/ToolsPopover';
 
 interface InputAreaProps {
   value: string;
@@ -409,572 +414,4 @@ export function InputArea({
   );
 }
 
-interface AccessoryButtonProps {
-  label: string;
-  displayLabel?: string;
-  icon?: string;
-  onClick: () => void;
-  badge?: number;
-  comingSoon?: boolean;
-}
-
-interface LlmProfileSelectorProps {
-  profileId: string | null;
-  profiles: string[];
-  onSelect: (profileId: string) => void;
-  onOpenCreate?: () => void;
-  onOpenEdit?: (profileId: string) => void;
-}
-
-function LlmProfileSelector({
-  profileId,
-  profiles,
-  onSelect,
-  onOpenCreate,
-  onOpenEdit,
-}: LlmProfileSelectorProps) {
-  const popoverRef = useRef<HTMLDivElement>(null);
-  const [isOpen, setIsOpen] = useState(false);
-
-  useCloseOnEscapeAndOutsideClick({ isOpen, onClose: () => setIsOpen(false), ref: popoverRef, delay: 100 });
-
-  const sanitizedProfiles = profiles.filter((id) => typeof id === 'string' && id.trim().length > 0);
-  const hasProfiles = sanitizedProfiles.length > 0;
-  const hasValidSelection = typeof profileId === 'string' && sanitizedProfiles.includes(profileId);
-  const selectedProfileId = hasValidSelection ? profileId : null;
-  const shouldPromptCreate = selectedProfileId === null && !hasProfiles && Boolean(onOpenCreate);
-  const shown = selectedProfileId ?? (hasProfiles ? 'Select profile…' : 'New profile…');
-  const tooltip = `LLM profile: ${shown}`;
-
-  const handleSelect = (next: string) => {
-    onSelect(next);
-    setIsOpen(false);
-  };
-
-  const isSelected = (candidate: string) => candidate === selectedProfileId;
-
-  return (
-    <div className="relative">
-      <button
-        type="button"
-        onClick={() => {
-          if (shouldPromptCreate && onOpenCreate) {
-            onOpenCreate();
-            return;
-          }
-          setIsOpen((prev) => !prev);
-        }}
-        className={`
-          inline-flex items-center gap-2
-          px-3 py-2 rounded-lg
-          text-xs font-medium
-          transition-all duration-200
-          border
-          focus:outline-none focus:ring-2 focus:ring-brand-500/40 focus:ring-offset-0
-          bg-white/[0.04] text-stone-400 border-white/[0.06]
-          hover:bg-white/[0.08] hover:text-stone-300 hover:border-white/[0.1]
-        `}
-        aria-label="LLM profile"
-        title={tooltip}
-      >
-        <span className="codicon codicon-symbol-parameter text-[13px] text-brand-400/70" />
-        <span className="font-mono text-stone-300 truncate max-w-[14rem]">{shown}</span>
-        <span className={`codicon codicon-chevron-${isOpen ? 'up' : 'down'} text-[12px] opacity-70`} />
-      </button>
-
-      {isOpen && (
-        <div
-          ref={popoverRef}
-          className="absolute bottom-full left-0 mb-2 w-72 bg-[var(--vscode-editor-background)] border border-white/10 rounded-xl shadow-2xl overflow-hidden animate-slide-up z-50"
-        >
-          <div className="px-4 py-3 border-b border-white/10">
-            <div className="flex items-center gap-2">
-              <span className="codicon codicon-symbol-parameter text-brand-400" />
-              <h3 className="font-semibold text-sm">LLM Profile</h3>
-            </div>
-            </div>
-
-            <div className="p-2 space-y-1" role="listbox" aria-label="LLM profiles">
-              {sanitizedProfiles.length === 0 ? (
-                <div className="px-3 py-2 text-sm opacity-60">No profiles found</div>
-              ) : (
-                sanitizedProfiles.map((id) => {
-                  const selected = isSelected(id);
-                return (
-                  <div
-                    key={id}
-                    role="option"
-                    aria-selected={selected}
-                    className={`
-                      w-full text-left px-3 py-2 rounded-lg
-                      text-sm
-                      transition-colors duration-150
-                      hover:bg-white/10
-                      flex items-center gap-2
-                      ${selected ? 'bg-brand-500/20 text-brand-300' : 'text-stone-300'}
-                    `}
-                  >
-                    <button
-                      type="button"
-                      onClick={() => handleSelect(id)}
-                      className="flex-1 flex items-center gap-2 text-left min-w-0"
-                      aria-label={`Select profile ${id}`}
-                      title={`Select profile ${id}`}
-                    >
-                      <span className="codicon codicon-symbol-misc" />
-                      <span className="flex-1 font-mono truncate">{id}</span>
-                    </button>
-
-                    {selected && onOpenEdit && (
-                      <button
-                        type="button"
-                        onClick={() => {
-                          setIsOpen(false);
-                          onOpenEdit(id);
-                        }}
-                        className="h-7 w-7 rounded-md bg-white/[0.03] border border-white/[0.06] text-stone-300 hover:bg-white/[0.08] hover:border-white/[0.1] transition-all flex items-center justify-center"
-                        aria-label={`Edit selected profile ${id}`}
-                        title={`Edit profile ${id}`}
-                      >
-                        <span className="codicon codicon-settings-gear text-[13px]" />
-                      </button>
-                    )}
-
-                    {selected && <span className="codicon codicon-check text-brand-400" />}
-                  </div>
-                  );
-                })
-              )}
-
-              <div className="my-1 border-t border-white/10" />
-
-              <button
-                type="button"
-                onClick={() => {
-                if (!onOpenCreate) return;
-                setIsOpen(false);
-                onOpenCreate();
-              }}
-              className={`
-                w-full text-left px-3 py-2 rounded-lg
-                text-sm
-                transition-colors duration-150
-                hover:bg-white/10
-                flex items-center gap-2
-                ${onOpenCreate ? 'text-stone-300' : 'text-stone-500 cursor-not-allowed'}
-              `}
-              disabled={!onOpenCreate}
-              aria-label="New profile…"
-              title="New profile…"
-            >
-              <span className="codicon codicon-add" />
-              <span className="flex-1">New profile…</span>
-            </button>
-          </div>
-        </div>
-      )}
-    </div>
-  );
-}
-
-function AccessoryButton({ icon, label, displayLabel, onClick, badge, comingSoon }: AccessoryButtonProps) {
-  return (
-    <button
-      onClick={onClick}
-      disabled={comingSoon}
-      className={`
-        relative inline-flex items-center gap-2
-        px-3 py-2 rounded-lg
-        text-xs font-medium
-        transition-all duration-200
-        border
-        focus:outline-none focus:ring-2 focus:ring-brand-500/40 focus:ring-offset-0
-        ${comingSoon
-          ? 'bg-white/[0.02] text-stone-600 border-white/[0.03] cursor-not-allowed'
-          : 'bg-white/[0.04] text-stone-400 border-white/[0.06] hover:bg-white/[0.08] hover:text-stone-300 hover:border-white/[0.1]'
-        }
-      `}
-      aria-label={label}
-      title={label}
-    >
-      {icon && <span className={`codicon codicon-${icon}`} />}
-      <span>{displayLabel ?? label}</span>
-
-      {badge !== undefined && badge > 0 && (
-        <span className="absolute -top-1.5 -right-1.5 h-4 min-w-4 px-1 rounded-full bg-gradient-to-b from-brand-400 to-brand-600 text-white text-[10px] font-semibold flex items-center justify-center shadow-glow-sm">
-          {badge}
-        </span>
-      )}
-
-      {comingSoon && (
-        <span className="text-[10px] text-stone-600 italic">soon</span>
-      )}
-    </button>
-  );
-}
-
-{/* Context Picker Popover */}
-interface ContextPickerProps {
-  isOpen: boolean;
-  onClose: (reason: CloseReason) => void;
-  files: string[];
-  selectedFiles: string[];
-  onToggleFile: (file: string) => void;
-  searchQuery: string;
-  onSearchChange: (query: string) => void;
-}
-
-export function ContextPicker({
-  isOpen,
-  onClose,
-  files,
-  selectedFiles,
-  onToggleFile,
-  searchQuery,
-  onSearchChange,
-}: ContextPickerProps) {
-  const popoverRef = useRef<HTMLDivElement>(null);
-  const [activeIndex, setActiveIndex] = useState(0);
-
-  useCloseOnEscapeAndOutsideClick({ isOpen, onClose, ref: popoverRef, delay: 100 });
-
-  const filteredFiles = files.filter((file) =>
-    file.toLowerCase().includes(searchQuery.toLowerCase())
-  );
-
-  const listboxId = 'context-picker-listbox';
-  const safeActiveIndex = filteredFiles.length > 0 ? Math.min(activeIndex, filteredFiles.length - 1) : 0;
-  const activeOptionId = filteredFiles.length > 0 ? `context-picker-option-${safeActiveIndex}` : undefined;
-
-  const handleSearchKeyDown = (e: React.KeyboardEvent<HTMLInputElement>) => {
-    if (e.key === 'ArrowDown') {
-      e.preventDefault();
-      setActiveIndex(() => Math.min(safeActiveIndex + 1, Math.max(filteredFiles.length - 1, 0)));
-      return;
-    }
-
-    if (e.key === 'ArrowUp') {
-      e.preventDefault();
-      setActiveIndex(() => Math.max(safeActiveIndex - 1, 0));
-      return;
-    }
-
-    if (e.key === 'Enter') {
-      const file = filteredFiles[safeActiveIndex];
-      if (!file) return;
-      e.preventDefault();
-      onToggleFile(file);
-      return;
-    }
-
-    if (e.key === 'Escape') {
-      e.preventDefault();
-      onClose('escape');
-    }
-  };
-
-  if (!isOpen) return null;
-
-  return (
-    <div
-      ref={popoverRef}
-      className="absolute bottom-full left-0 mb-1 w-80 max-h-96 bg-[var(--vscode-editor-background)] border border-white/[0.08] rounded-xl shadow-2xl overflow-hidden animate-slide-up z-50"
-      style={{
-        background: 'linear-gradient(135deg, rgba(28, 25, 23, 0.98) 0%, rgba(12, 10, 9, 0.98) 100%)',
-      }}
-    >
-      {/* Header */}
-      <div className="px-4 py-3 border-b border-white/[0.06]">
-        <div className="flex items-center gap-2 mb-2">
-          <span className="codicon codicon-mention text-brand-400" />
-          <h3 className="font-semibold text-sm text-stone-200">Add Context Files</h3>
-        </div>
-        <input
-          type="text"
-          value={searchQuery}
-          onChange={(e) => {
-            onSearchChange(e.target.value);
-            setActiveIndex(0);
-          }}
-          onKeyDown={handleSearchKeyDown}
-          placeholder="Search files..."
-          aria-controls={listboxId}
-          aria-activedescendant={activeOptionId}
-          className="w-full px-3 py-2 bg-black/30 border border-white/[0.08] rounded-lg text-sm text-stone-200 placeholder:text-stone-500 focus:outline-none focus:ring-2 focus:ring-brand-500/40 focus:border-brand-500/30"
-          autoFocus
-        />
-      </div>
-
-      {/* File list */}
-      <div className="overflow-y-auto max-h-64">
-        {filteredFiles.length === 0 ? (
-          <div className="px-4 py-8 text-center text-sm text-stone-500">
-            No matches
-          </div>
-        ) : (
-          <div className="p-2 space-y-0.5" role="listbox" id={listboxId}>
-            {filteredFiles.map((file, index) => {
-              const isSelected = selectedFiles.includes(file);
-              const isActive = index === safeActiveIndex;
-              return (
-                <button
-                  key={file}
-                  onClick={() => onToggleFile(file)}
-                  role="option"
-                  id={`context-picker-option-${index}`}
-                  aria-label={file}
-                  aria-selected={isSelected ? 'true' : 'false'}
-                  className={`
-                    w-full text-left px-3 py-2 rounded-lg
-                    text-sm font-mono
-                    transition-all duration-150
-                    flex items-center gap-2
-                    ${isSelected
-                      ? 'bg-brand-500/15 text-brand-300 border border-brand-500/20'
-                      : isActive
-                        ? 'bg-white/[0.06] text-stone-200 border border-white/[0.16]'
-                      : 'text-stone-400 hover:bg-white/[0.04] hover:text-stone-300 border border-transparent'
-                    }
-                  `}
-                >
-                  <span className={`codicon codicon-${isSelected ? 'check' : 'file'} ${isSelected ? 'text-brand-400' : 'text-stone-500'}`} />
-                  <span className="truncate flex-1">{file}</span>
-                </button>
-              );
-            })}
-          </div>
-        )}
-      </div>
-
-      {/* Footer */}
-      {selectedFiles.length > 0 && (
-        <div className="px-4 py-2 border-t border-white/[0.06] bg-brand-500/5 text-xs text-brand-300">
-          {selectedFiles.length} file{selectedFiles.length !== 1 ? 's' : ''} selected
-        </div>
-      )}
-    </div>
-  );
-}
-
-// Tools Popover
-interface ToolDescriptor {
-  id: string;
-  label: string;
-}
-
-interface ToolsPopoverProps {
-  isOpen: boolean;
-  onClose: (reason: CloseReason) => void;
-  tools: ToolDescriptor[];
-  enabledToolIds: string[];
-  onToggleTool: (toolId: string) => void;
-}
-
-export function ToolsPopover({
-  isOpen,
-  onClose,
-  tools,
-  enabledToolIds,
-  onToggleTool,
-}: ToolsPopoverProps) {
-  const popoverRef = useRef<HTMLDivElement>(null);
-
-  useCloseOnEscapeAndOutsideClick({ isOpen, onClose, ref: popoverRef, delay: 100 });
-
-  if (!isOpen) return null;
-
-  return (
-    <div
-      ref={popoverRef}
-      className="absolute bottom-full left-0 mb-1 w-64 max-h-96 bg-[var(--vscode-editor-background)] border border-white/[0.08] rounded-xl shadow-2xl overflow-hidden animate-slide-up z-50"
-      style={{
-        background: 'linear-gradient(135deg, rgba(28, 25, 23, 0.98) 0%, rgba(12, 10, 9, 0.98) 100%)',
-      }}
-    >
-      <div className="px-4 py-3 border-b border-white/[0.06]">
-        <div className="flex items-center gap-2">
-          <span className="codicon codicon-tools text-brand-400" />
-          <h3 className="font-semibold text-sm text-stone-200">Tools</h3>
-        </div>
-        <div className="mt-1 text-xs text-stone-500">
-          Select which tools the agent can use
-        </div>
-      </div>
-
-      <div className="overflow-y-auto max-h-64">
-        {tools.length === 0 ? (
-          <div className="px-4 py-8 text-center text-sm text-stone-500">
-            No tools available
-          </div>
-        ) : (
-          <div className="p-2 space-y-0.5" role="listbox" aria-label="Tools">
-            {tools.map((tool) => {
-              const isEnabled = enabledToolIds.includes(tool.id);
-              return (
-                <button
-                  key={tool.id}
-                  type="button"
-                  onClick={() => onToggleTool(tool.id)}
-                  role="option"
-                  aria-label={tool.label}
-                  aria-selected={isEnabled ? 'true' : 'false'}
-                  className={`
-                    w-full text-left px-3 py-2 rounded-lg
-                    text-sm
-                    transition-all duration-150
-                    flex items-center gap-2
-                    group
-                    ${isEnabled
-                      ? 'bg-brand-500/10 text-stone-200 border border-brand-500/25'
-                      : 'text-stone-400 hover:bg-white/[0.04] hover:text-stone-300 border border-transparent'
-                    }
-                  `}
-                >
-                  <span className="codicon codicon-symbol-method text-brand-400/70" />
-                  <span className="flex-1 truncate">{tool.label}</span>
-                  {isEnabled && <span className="codicon codicon-check text-brand-400" />}
-                </button>
-              );
-            })}
-          </div>
-        )}
-      </div>
-    </div>
-  );
-}
-
-// Skills Popover
-interface Skill {
-  label: string;
-  path: string;
-}
-
-interface SkillsPopoverProps {
-  isOpen: boolean;
-  onClose: (reason: CloseReason) => void;
-  skills: Skill[];
-  onOpenSkill: (path: string) => void;
-}
-
-export function SkillsPopover({
-  isOpen,
-  onClose,
-  skills,
-  onOpenSkill,
-}: SkillsPopoverProps) {
-  const popoverRef = useRef<HTMLDivElement>(null);
-  const [searchQuery, setSearchQuery] = useState('');
-  const [activeIndex, setActiveIndex] = useState(0);
-
-  useCloseOnEscapeAndOutsideClick({ isOpen, onClose, ref: popoverRef, delay: 100 });
-
-  const filteredSkills = skills.filter((skill) =>
-    skill.label.toLowerCase().includes(searchQuery.toLowerCase())
-  );
-
-  const listboxId = 'skills-picker-listbox';
-  const safeActiveIndex = filteredSkills.length > 0 ? Math.min(activeIndex, filteredSkills.length - 1) : 0;
-  const activeOptionId = filteredSkills.length > 0 ? `skills-picker-option-${safeActiveIndex}` : undefined;
-
-  const handleSearchKeyDown = (e: React.KeyboardEvent<HTMLInputElement>) => {
-    if (e.key === 'ArrowDown') {
-      e.preventDefault();
-      setActiveIndex(() => Math.min(safeActiveIndex + 1, Math.max(filteredSkills.length - 1, 0)));
-      return;
-    }
-
-    if (e.key === 'ArrowUp') {
-      e.preventDefault();
-      setActiveIndex(() => Math.max(safeActiveIndex - 1, 0));
-      return;
-    }
-
-    if (e.key === 'Enter') {
-      const skill = filteredSkills[safeActiveIndex];
-      if (!skill) return;
-      e.preventDefault();
-      onOpenSkill(skill.path);
-      return;
-    }
-
-    if (e.key === 'Escape') {
-      e.preventDefault();
-      onClose('escape');
-    }
-  };
-
-  if (!isOpen) return null;
-
-  return (
-    <div
-      ref={popoverRef}
-      className="absolute bottom-full left-0 mb-1 w-64 max-h-96 bg-[var(--vscode-editor-background)] border border-white/[0.08] rounded-xl shadow-2xl overflow-hidden animate-slide-up z-50"
-      style={{
-        background: 'linear-gradient(135deg, rgba(28, 25, 23, 0.98) 0%, rgba(12, 10, 9, 0.98) 100%)',
-      }}
-    >
-      {/* Header */}
-      <div className="px-4 py-3 border-b border-white/[0.06]">
-        <div className="flex items-center gap-2">
-          <span className="codicon codicon-mortar-board text-brand-400" />
-          <h3 className="font-semibold text-sm text-stone-200">Available Skills</h3>
-        </div>
-        <input
-          type="text"
-          value={searchQuery}
-          onChange={(e) => {
-            setSearchQuery(e.target.value);
-            setActiveIndex(0);
-          }}
-          onKeyDown={handleSearchKeyDown}
-          placeholder="Search skills..."
-          aria-controls={listboxId}
-          aria-activedescendant={activeOptionId}
-          className="mt-2 w-full px-3 py-2 bg-black/30 border border-white/[0.08] rounded-lg text-sm text-stone-200 placeholder:text-stone-500 focus:outline-none focus:ring-2 focus:ring-brand-500/40 focus:border-brand-500/30"
-          autoFocus
-        />
-      </div>
-
-      {/* Skills list */}
-      <div className="overflow-y-auto max-h-64">
-        {filteredSkills.length === 0 ? (
-          <div className="px-4 py-8 text-center text-sm text-stone-500">
-            No skills found
-          </div>
-        ) : (
-          <div className="p-2 space-y-0.5" role="listbox" aria-label="Skills" id={listboxId}>
-            {filteredSkills.map((skill, index) => {
-              const isActive = index === safeActiveIndex;
-              return (
-              <button
-                key={skill.path}
-                onClick={() => onOpenSkill(skill.path)}
-                role="option"
-                id={`skills-picker-option-${index}`}
-                aria-label={skill.label}
-                aria-selected="false"
-                className={`
-                  w-full text-left px-3 py-2 rounded-lg
-                  text-sm
-                  transition-all duration-150
-                  flex items-center gap-2
-                  group
-                  ${isActive
-                    ? 'bg-brand-500/10 text-stone-200 border border-brand-500/25'
-                    : 'text-stone-400 hover:bg-white/[0.04] hover:text-stone-300 border border-transparent'
-                  }
-                `}
-                data-active={isActive ? 'true' : 'false'}
-              >
-                <span className="codicon codicon-file-code text-brand-400/70" />
-                <span className="flex-1 truncate">{skill.label}</span>
-                <span className="codicon codicon-arrow-right text-stone-600 group-hover:text-stone-400 transition-colors" />
-              </button>
-              );
-            })}
-          </div>
-        )}
-      </div>
-    </div>
-  );
-}
+export { ContextPicker, SkillsPopover, ToolsPopover };
